Issue #2680 Fixed various OpenGL errors
(cherry picked from commit 061de63c9ba76794041cc926db01cbbd1079f511 [formerly 5d488d27b9d7b7035f7d4173ee816717378eef02] [formerly 83c600ada9a9b8a70757bccd1b6af023d1637b51] [formerly 431ffa97ae4b1c9159c4607e51dbed23c77b76d2 [formerly f4765d157724f4cb4a58a84c773cab5ec12c8c5e] [formerly 83c600ada9a9b8a70757bccd1b6af023d1637b51 [formerly 59d4c2b9965a7c208863af8f5a3ba4145ea2bbf4]]])

Former-commit-id: 787e48e9a7b4803ff39eafa6ffcaf58ac13408a7 [formerly d566ed99276e94286271206981916244a6201255] [formerly 0c840cf724f11c5f6e83124ccb070cdfd442aa2e [formerly 6c3817256b4cf6b5deca8d73d0df1d1f89b4ea82]]
Former-commit-id: 0c840cf724f11c5f6e83124ccb070cdfd442aa2e
Former-commit-id: 122c3442f80fe851ea4ddfbd121cbbf68dc5276a

+++ cave/com.raytheon.viz.core.gl/localization/glsl/include/mapping.glsl
@@ -203,7 +203,7 @@ float valueToLogIndex(float value, float rangeMin, float rangeMax) {
 		return 0;
 	}
 
-	double index = (log(value) - log(rangeMin))
+	float index = (log(value) - log(rangeMin))
 			/ (log(rangeMax) - log(rangeMin));
 	if (reverse != 0) {
 		index = -index;
